add .net standard project and modify some denpendcy
