+++ GoogleAnalyticsClientDotNet.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29230.47
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "App", "App", "{CE66EEDC-6C40-4821-9566-AC5D782CEA31}"
 EndProject
@@ -19,9 +19,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test", "Test", "{52D1D9E1-2
 EndProject
 Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "GoogleAnalyticsClientDotNet.Shared", "Library\GoogleAnalyticsClientDotNet.Shared\GoogleAnalyticsClientDotNet.Shared.shproj", "{C47FD564-E9D8-4042-BA81-8EBF448D3BBE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoogleAnalyticsClientDotNet.DotNetStandard", "Library\GoogleAnalyticsClientDotNet.DotNetStandard\GoogleAnalyticsClientDotNet.DotNetStandard.csproj", "{D02BDC37-4D09-404C-B884-2A158EBC2353}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BackgroundTask", "BackgroundTask", "{95353EC2-94E3-493A-80F8-2F25473EC07F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoogleAnalyticsClientDotNet.Standard", "Library\GoogleAnalyticsClientDotNet.Standard\GoogleAnalyticsClientDotNet.Standard.csproj", "{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
@@ -108,22 +108,22 @@ Global
 		{3BC2114D-A1F5-4AA6-98E8-E39FC5217224}.Release|x64.Build.0 = Release|Any CPU
 		{3BC2114D-A1F5-4AA6-98E8-E39FC5217224}.Release|x86.ActiveCfg = Release|Any CPU
 		{3BC2114D-A1F5-4AA6-98E8-E39FC5217224}.Release|x86.Build.0 = Release|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Debug|ARM.Build.0 = Debug|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Debug|x64.Build.0 = Debug|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Debug|x86.Build.0 = Debug|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Release|ARM.ActiveCfg = Release|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Release|ARM.Build.0 = Release|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Release|x64.ActiveCfg = Release|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Release|x64.Build.0 = Release|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Release|x86.ActiveCfg = Release|Any CPU
-		{D02BDC37-4D09-404C-B884-2A158EBC2353}.Release|x86.Build.0 = Release|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Debug|ARM.Build.0 = Debug|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Debug|x64.Build.0 = Debug|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Debug|x86.Build.0 = Debug|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Release|ARM.ActiveCfg = Release|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Release|ARM.Build.0 = Release|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Release|x64.ActiveCfg = Release|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Release|x64.Build.0 = Release|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Release|x86.ActiveCfg = Release|Any CPU
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -134,6 +134,9 @@ Global
 		{90D8D711-16EA-48AF-819E-3766F34FA864} = {CE66EEDC-6C40-4821-9566-AC5D782CEA31}
 		{3BC2114D-A1F5-4AA6-98E8-E39FC5217224} = {9FA35DD2-49B9-4D87-90C7-60AF1BD15037}
 		{C47FD564-E9D8-4042-BA81-8EBF448D3BBE} = {9FA35DD2-49B9-4D87-90C7-60AF1BD15037}
-		{D02BDC37-4D09-404C-B884-2A158EBC2353} = {9FA35DD2-49B9-4D87-90C7-60AF1BD15037}
+		{D352D8DE-7202-4B9D-8D9F-B4533B66F91A} = {9FA35DD2-49B9-4D87-90C7-60AF1BD15037}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {199A812C-BC8D-48CA-B061-0AAFEBDAED9E}
 	EndGlobalSection
 EndGlobal
