--- Library/GoogleAnalyticsClientDotNet.DotNetStandard/project.json
@@ -1,10 +1,0 @@
-{
-  "supports": {},
-  "dependencies": {
-    "NETStandard.Library": "1.6.0",
-    "Microsoft.NETCore.Portable.Compatibility": "1.0.1"
-  },
-  "frameworks": {
-    "netstandard1.2": {}
-  }
-}
