+++ Library/GoogleAnalyticsClientDotNet.Net45/FileLocalTracker.Net45.cs
@@ -8,8 +8,20 @@ using System.Threading.Tasks;
 
 namespace GoogleAnalyticsClientDotNet
 {
-    public partial class FileLocalTracker : ILocalTracker
+    public class FileLocalTracker : ILocalTracker
     {
+        public string SourceName { get; set; } = "default_ga_track_file.mtf";
+
+        public async Task<IEnumerable<string>> ReadTrackAsync()
+        {
+            return await ReadLocalFileAsync();
+        }
+
+        public async Task WriteTracksAsync(IEnumerable<string> tracks, bool replace = false)
+        {
+            await WriteLocalFileAsync(tracks, replace);
+        }
+
         private Task<IEnumerable<string>> ReadLocalFileAsync()
         {
             IEnumerable<string> content = null;
